+++ src/redux/actions/index.js
@@ -52,4 +52,5 @@ export * from './usage-consolidation';
 export * from './currencies';
 export * from './costperuse';
 export * from './package-titles';
+export * from './provider-packages';
 export * from './uc-credentials';