+++ src/redux/actions/provider-packages/clear-provider-packages.js
@@ -1,0 +1,5 @@
+export const CLEAR_PROVIDER_PACKAGES = 'CLEAR_PROVIDER_PACKAGES';
+
+export function clearProviderPackages() {
+  return { type: CLEAR_PROVIDER_PACKAGES };
+}