+++ src/redux/actions/provider-packages/get-provider-packages-success.js
@@ -1,0 +1,11 @@
+export const GET_PROVIDER_PACKAGES_SUCCESS = 'GET_PROVIDER_PACKAGES_SUCCESS';
+
+export function getProviderPackagesSuccess({ data, meta }) {
+  return {
+    type: GET_PROVIDER_PACKAGES_SUCCESS,
+    payload: {
+      data,
+      totalResults: meta.totalResults,
+    },
+  };
+}