+++ src/redux/reducers/index.js
@@ -10,4 +10,5 @@ export { default as usageConsolidation } from './usageConsolidation';
 export { default as currencies } from './currencies';
 export { default as costPerUse } from './costPerUse';
 export { default as packageTitles } from './packageTitles';
+export { default as providerPackages } from './providerPackages';
 export { default as ucCredentials } from './ucCredentials';